disable old png renderer if using cairo
