+++ graphutil.go
@@ -1,0 +1,150 @@
+package main
+
+import (
+	"image/color"
+	"strconv"
+)
+
+func getBool(s string, def bool) bool {
+	if s == "" {
+		return def
+	}
+
+	switch s {
+	case "True", "true", "1":
+		return true
+	case "False", "false", "0":
+		return false
+	}
+
+	return def
+}
+
+func getString(s string, def string) string {
+	if s == "" {
+		return def
+	}
+
+	return s
+}
+
+func getFloat64(s string, def float64) float64 {
+	if s == "" {
+		return def
+	}
+
+	n, err := strconv.ParseFloat(s, 64)
+	if err != nil {
+		return def
+	}
+
+	return float64(n)
+}
+
+var colors = map[string]color.RGBA{
+	// Graphite default colors
+	"black":     color.RGBA{0x00, 0x00, 0x00, 0xff},
+	"white":     color.RGBA{0xff, 0xff, 0xff, 0xff},
+	"blue":      color.RGBA{0x64, 0x64, 0xff, 0xff},
+	"green":     color.RGBA{0x00, 0xc8, 0x00, 0xff},
+	"red":       color.RGBA{0xc8, 0x00, 0x32, 0xff},
+	"yellow":    color.RGBA{0xff, 0xff, 0x00, 0xff},
+	"orange":    color.RGBA{0xff, 0xa5, 0x00, 0xff},
+	"purple":    color.RGBA{0xc8, 0x64, 0xff, 0xff},
+	"brown":     color.RGBA{0x96, 0x64, 0x32, 0xff},
+	"cyan":      color.RGBA{0x00, 0xff, 0xff, 0xff},
+	"aqua":      color.RGBA{0x00, 0x96, 0x96, 0xff},
+	"gray":      color.RGBA{0xaf, 0xaf, 0xaf, 0xff},
+	"grey":      color.RGBA{0xaf, 0xaf, 0xaf, 0xff},
+	"magenta":   color.RGBA{0xff, 0x00, 0xff, 0xff},
+	"pink":      color.RGBA{0xff, 0x64, 0x64, 0xff},
+	"gold":      color.RGBA{0xc8, 0xc8, 0x00, 0xff},
+	"rose":      color.RGBA{0xc8, 0x96, 0xc8, 0xff},
+	"darkblue":  color.RGBA{0x00, 0x00, 0xff, 0xff},
+	"darkgreen": color.RGBA{0x00, 0xff, 0x00, 0xff},
+	"darkred":   color.RGBA{0xff, 0x00, 0x00, 0xff},
+	"darkgray":  color.RGBA{0x6f, 0x6f, 0x6f, 0xff},
+	"darkgrey":  color.RGBA{0x6f, 0x6f, 0x6f, 0xff},
+
+	// Custom colors
+	"navy":                 color.RGBA{0x00, 0x00, 0x80, 0xff},
+	"mediumblue":           color.RGBA{0x00, 0x00, 0xcd, 0xff},
+	"teal":                 color.RGBA{0x00, 0x80, 0x80, 0xff},
+	"darkcyan":             color.RGBA{0x00, 0x8b, 0x8b, 0xff},
+	"deepskyblue":          color.RGBA{0x00, 0xbf, 0xff, 0xff},
+	"darkturquoise":        color.RGBA{0x00, 0xce, 0xd1, 0xff},
+	"mediumspringgreen":    color.RGBA{0x00, 0xfa, 0x9a, 0xff},
+	"lime":                 color.RGBA{0x00, 0xff, 0x00, 0xff},
+	"springgreen":          color.RGBA{0x00, 0xff, 0x7f, 0xff},
+	"midnightblue":         color.RGBA{0x19, 0x19, 0x70, 0xff},
+	"dodgerblue":           color.RGBA{0x1e, 0x90, 0xff, 0xff},
+	"lightseagreen":        color.RGBA{0x20, 0xb2, 0xaa, 0xff},
+	"forestgreen":          color.RGBA{0x22, 0x8b, 0x22, 0xff},
+	"seagreen":             color.RGBA{0x2e, 0x8b, 0x57, 0xff},
+	"darkslategray":        color.RGBA{0x2f, 0x4f, 0x4f, 0xff},
+	"limegreen":            color.RGBA{0x32, 0xcd, 0x32, 0xff},
+	"mediumseagreen":       color.RGBA{0x3c, 0xb3, 0x71, 0xff},
+	"turquoise":            color.RGBA{0x40, 0xe0, 0xd0, 0xff},
+	"royalblue":            color.RGBA{0x41, 0x69, 0xe1, 0xff},
+	"steelblue":            color.RGBA{0x46, 0x82, 0xb4, 0xff},
+	"darkslateblue":        color.RGBA{0x48, 0x3d, 0x8b, 0xff},
+	"mediumturquoise":      color.RGBA{0x48, 0xd1, 0xcc, 0xff},
+	"indigo":               color.RGBA{0x4b, 0x00, 0x82, 0xff},
+	"darkolivegreen":       color.RGBA{0x55, 0x6b, 0x2f, 0xff},
+	"cadetblue":            color.RGBA{0x5f, 0x9e, 0xa0, 0xff},
+	"cornflowerblue":       color.RGBA{0x64, 0x95, 0xed, 0xff},
+	"mediumaquamarine":     color.RGBA{0x66, 0xcd, 0xaa, 0xff},
+	"dimgray":              color.RGBA{0x69, 0x69, 0x69, 0xff},
+	"slateblue":            color.RGBA{0x6a, 0x5a, 0xcd, 0xff},
+	"olivedrab":            color.RGBA{0x6b, 0x8e, 0x23, 0xff},
+	"slategray":            color.RGBA{0x70, 0x80, 0x90, 0xff},
+	"lightslategray":       color.RGBA{0x77, 0x88, 0x99, 0xff},
+	"mediumslateblue":      color.RGBA{0x7b, 0x68, 0xee, 0xff},
+	"lawngreen":            color.RGBA{0x7c, 0xfc, 0x00, 0xff},
+	"chartreuse":           color.RGBA{0x7f, 0xff, 0x00, 0xff},
+	"aquamarine":           color.RGBA{0x7f, 0xff, 0xd4, 0xff},
+	"lavender":             color.RGBA{0xe6, 0xe6, 0xfa, 0xff},
+	"darksalmon":           color.RGBA{0xe9, 0x96, 0x7a, 0xff},
+	"violet":               color.RGBA{0xee, 0x82, 0xee, 0xff},
+	"palegoldenrod":        color.RGBA{0xee, 0xe8, 0xaa, 0xff},
+	"lightcoral":           color.RGBA{0xf0, 0x80, 0x80, 0xff},
+	"khaki":                color.RGBA{0xf0, 0xe6, 0x8c, 0xff},
+	"aliceblue":            color.RGBA{0xf0, 0xf8, 0xff, 0xff},
+	"honeydew":             color.RGBA{0xf0, 0xff, 0xf0, 0xff},
+	"azure":                color.RGBA{0xf0, 0xff, 0xff, 0xff},
+	"sandybrown":           color.RGBA{0xf4, 0xa4, 0x60, 0xff},
+	"wheat":                color.RGBA{0xf5, 0xde, 0xb3, 0xff},
+	"beige":                color.RGBA{0xf5, 0xf5, 0xdc, 0xff},
+	"whitesmoke":           color.RGBA{0xf5, 0xf5, 0xf5, 0xff},
+	"mintcream":            color.RGBA{0xf5, 0xff, 0xfa, 0xff},
+	"ghostwhite":           color.RGBA{0xf8, 0xf8, 0xff, 0xff},
+	"salmon":               color.RGBA{0xfa, 0x80, 0x72, 0xff},
+	"antiquewhite":         color.RGBA{0xfa, 0xeb, 0xd7, 0xff},
+	"linen":                color.RGBA{0xfa, 0xf0, 0xe6, 0xff},
+	"lightgoldenrodyellow": color.RGBA{0xfa, 0xfa, 0xd2, 0xff},
+	"oldlace":              color.RGBA{0xfd, 0xf5, 0xe6, 0xff},
+	"fuchsia":              color.RGBA{0xff, 0x00, 0xff, 0xff},
+	"deeppink":             color.RGBA{0xff, 0x14, 0x93, 0xff},
+	"orangered":            color.RGBA{0xff, 0x45, 0x00, 0xff},
+	"tomato":               color.RGBA{0xff, 0x63, 0x47, 0xff},
+	"hotpink":              color.RGBA{0xff, 0x69, 0xb4, 0xff},
+	"coral":                color.RGBA{0xff, 0x7f, 0x50, 0xff},
+	"darkorange":           color.RGBA{0xff, 0x8c, 0x00, 0xff},
+	"lightsalmon":          color.RGBA{0xff, 0xa0, 0x7a, 0xff},
+	"lightpink":            color.RGBA{0xff, 0xb6, 0xc1, 0xff},
+	"peachpuff":            color.RGBA{0xff, 0xda, 0xb9, 0xff},
+	"navajowhite":          color.RGBA{0xff, 0xde, 0xad, 0xff},
+	"moccasin":             color.RGBA{0xff, 0xe4, 0xb5, 0xff},
+	"bisque":               color.RGBA{0xff, 0xe4, 0xc4, 0xff},
+	"mistyrose":            color.RGBA{0xff, 0xe4, 0xe1, 0xff},
+	"blanchedalmond":       color.RGBA{0xff, 0xeb, 0xcd, 0xff},
+	"papayawhip":           color.RGBA{0xff, 0xef, 0xd5, 0xff},
+	"lavenderblush":        color.RGBA{0xff, 0xf0, 0xf5, 0xff},
+	"seashell":             color.RGBA{0xff, 0xf5, 0xee, 0xff},
+	"cornsilk":             color.RGBA{0xff, 0xf8, 0xdc, 0xff},
+	"lemonchiffon":         color.RGBA{0xff, 0xfa, 0xcd, 0xff},
+	"floralwhite":          color.RGBA{0xff, 0xfa, 0xf0, 0xff},
+	"snow":                 color.RGBA{0xff, 0xfa, 0xfa, 0xff},
+	"lightyellow":          color.RGBA{0xff, 0xff, 0xe0, 0xff},
+	"ivory":                color.RGBA{0xff, 0xff, 0xf0, 0xff},
+}
